[Fix] Some small errors in S3M saver. [Imp] Default global volume is now written to pattern when converting to XM and is detected as an MPT extension in "Detect MPT hacks". [Ref] Smaller stuff.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@1255 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Moddoc.cpp
@@ -873,8 +873,8 @@ UINT CModDoc::ShowLog(LPCSTR lpszTitle, CWnd *parent)
 }
 
 
-UINT CModDoc::GetPlaybackMidiChannel(const ModInstrument *pIns, CHANNELINDEX nChn) const
-//--------------------------------------------------------------------------------------
+uint8 CModDoc::GetPlaybackMidiChannel(const ModInstrument *pIns, CHANNELINDEX nChn) const
+//---------------------------------------------------------------------------------------
 {
 	if(pIns->nMidiChannel == MidiMappedChannel)
 	{
@@ -1025,7 +1025,7 @@ UINT CModDoc::PlayNote(UINT note, INSTRUMENTINDEX nins, SAMPLEINDEX nsmp, bool p
 				if ((!nPlugin || nPlugin > MAX_MIXPLUGINS) && nCurrentChn != CHANNELINDEX_INVALID)
 					nPlugin = m_SndFile.ChnSettings[nCurrentChn].nMixPlugin;	// Then try Channel VST
 				
-   				if ((nPlugin) && (nPlugin <= MAX_MIXPLUGINS))
+				if ((nPlugin) && (nPlugin <= MAX_MIXPLUGINS))
 				{
 					IMixPlugin *pPlugin = m_SndFile.m_MixPlugins[nPlugin - 1].pMixPlugin;
